% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{add_xx_yy}
\alias{add_xx_yy}
\alias{add_epci_of_com}
\alias{add_dep_of_com}
\alias{add_reg_of_com}
\title{Ajouter des variables sur le rattachement supra communal dans un dataframe a la commune}
\usage{
add_epci_of_com(.data, code_commune)

add_dep_of_com(.data, code_commune)

add_reg_of_com(.data, code_commune)
}
\arguments{
\item{.data}{Table en entrée.}

\item{code_commune}{Nom de la variable avec le code commune.}
}
\value{
un dataframe
}
\description{
Ajouter des variables sur le rattachement supra communal dans un dataframe a la commune

\code{add_epci_of_com()} ajouter le code epci

\code{add_dep_of_com()} ajouter le code département

\code{add_reg_of_com()} ajouter le code région
}
