% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arn_plm}
\alias{arn_plm}
\title{Table de passage des arrondissements de Paris, Lyon, Marseille vers leur code commune.}
\format{
A data frame with 45 rows and 2 variables:
\describe{
\item{ARN}{Code Insee de l'arrondissement}
\item{DEPCOM}{Code Insee de la commune}
}
}
\usage{
arn_plm
}
\description{
Un dataset contenant pour chaque arrondissement de Paris, Lyon et Marseille le code de la commune de rattachement
}
\keyword{datasets}
