% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_zone.R
\name{augmenter_liste_zone}
\alias{augmenter_liste_zone}
\title{augmenter_liste_zone() ajoute les lignes de totaux et sous totaux (France, province, DROM...) au df liste_zone}
\usage{
augmenter_liste_zone()
}
\value{
dataframe liste_zone avec ajout des lignes de total et sous total
}
\description{
augmenter_liste_zone() ajoute les lignes de totaux et sous totaux (France, province, DROM...) au df liste_zone
}
