% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_com_existe}
\alias{code_com_existe}
\title{teste que le code commune existe}
\usage{
code_com_existe(commune)
}
\arguments{
\item{commune}{Character. Le code de la commune concernée}
}
\value{
TRUE/FALSE
}
\description{
teste que le code commune existe
}
\examples{
code_com_existe("75056")
}
