% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_com_non_a_jour}
\alias{code_com_non_a_jour}
\title{test si le code commune est un code ayant existé mais qui n'est plus valide}
\usage{
code_com_non_a_jour(commune)
}
\arguments{
\item{commune}{Character. code commune}
}
\value{
TRUE/FALSE
}
\description{
test si le code commune est un code ayant existé mais qui n'est plus valide
}
\examples{
code_com_non_a_jour("75056")
}
