% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_com_plm}
\alias{code_com_plm}
\title{test si le code commune est celui d'un arrondissement de Paris, Lyon ou Marseille}
\usage{
code_com_plm(commune)
}
\arguments{
\item{commune}{Character. code commune}
}
\value{
TRUE/FALSE
}
\description{
test si le code commune est celui d'un arrondissement de Paris, Lyon ou Marseille
}
\examples{
code_com_plm("75120")
}
