% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_dep_existe}
\alias{code_dep_existe}
\title{teste que le code departement existe}
\usage{
code_dep_existe(departement)
}
\arguments{
\item{departement}{Character. Le code du departement concerné}
}
\value{
TRUE/FALSE
}
\description{
teste que le code departement existe
}
\examples{
code_dep_existe("24")
}
