% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_dep_of_com}
\alias{code_dep_of_com}
\title{Renvoie le code du département d appartenance d une commune}
\usage{
code_dep_of_com(com)
}
\arguments{
\item{com}{Character. Le code de la commune concernée}
}
\value{
un caractere
}
\description{
Renvoie le code du département d appartenance d une commune
}
\examples{
code_dep_of_com("24037")
}
