% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_epci_existe}
\alias{code_epci_existe}
\title{teste que le code EPCI existe}
\usage{
code_epci_existe(epci)
}
\arguments{
\item{epci}{Character. Le code de l'EPCI concerné}
}
\value{
TRUE/FALSE
}
\description{
teste que le code EPCI existe
}
\examples{
code_epci_existe("249740077")
}
