% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_reg_existe}
\alias{code_reg_existe}
\title{teste que le code region existe}
\usage{
code_reg_existe(region)
}
\arguments{
\item{region}{Character. Le code de la region concernee}
}
\value{
TRUE/FALSE
}
\description{
teste que le code region existe
}
\examples{
code_reg_existe("75")
}
