% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_reg_of_com}
\alias{code_reg_of_com}
\title{Renvoie le code région d appartenance d une commune}
\usage{
code_reg_of_com(depcom)
}
\arguments{
\item{depcom}{Character. Un vecteur de codes commune.}
}
\value{
Un vecteur.
}
\description{
Renvoie le code région d appartenance d une commune
}
\examples{
code_reg_of_com(c("24037","44109"))
}
