% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_reg_of_dep}
\alias{code_reg_of_dep}
\title{Renvoie les codes de la région d appartenance de chaque département}
\usage{
code_reg_of_dep(dep)
}
\arguments{
\item{dep}{Character. Un vecteur de codes département}
}
\value{
Un vecteur de caractère.
}
\description{
Renvoie les codes de la région d appartenance de chaque département
}
\examples{
code_reg_of_dep(c("75","44"))
}
