% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_reg_of_epci}
\alias{code_reg_of_epci}
\title{Renvoie le code du région d appartenance de chaque epci}
\usage{
code_reg_of_epci(epci)
}
\arguments{
\item{epci}{Character. Un vecteur de codes epci.}
}
\value{
une liste
}
\description{
Renvoie le code du région d appartenance de chaque epci
}
\examples{
code_reg_of_epci(c("200070647"))
}
