% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_xx_existe}
\alias{code_xx_existe}
\alias{code_reg_existe}
\alias{code_dep_existe}
\alias{code_epci_existe}
\alias{code_com_existe}
\title{teste que le code du territoire existe}
\usage{
code_reg_existe(reg)

code_dep_existe(dep)

code_epci_existe(epci)

code_com_existe(depcom)
}
\arguments{
\item{reg}{Vecteur de caractère. Le code de la region concernée.}

\item{dep}{Vecteur de caractère. Le code du departement concernée.}

\item{epci}{Vecteur de caractère. Le code de l'epci concernée.}

\item{depcom}{Vecteur de caractère. Le code de la commune concernée.}
}
\value{
Vecteur de booléen. TRUE si le code existe.
}
\description{
teste que le code du territoire existe

teste que des codes regions existent

teste que des codes département existent

Test que les codes EPCI existent

teste que les codes commune existent
}
\examples{
code_reg_existe(c("11", "99"))
code_dep_existe(c("75", "99"))
code_epci_existe("249740077")
code_com_existe("75056")
}
