% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_zone.R
\name{code_zone}
\alias{code_zone}
\title{Obtenir le code d'une zone}
\usage{
code_zone(type_zone, zone)
}
\arguments{
\item{type_zone}{Type de Zone}

\item{zone}{Nom de la Zone}
}
\value{
une chaîne de caractère
}
\description{
Obtenir le code d'une zone
}
\examples{
code_zone("Communes", "Nantes")
}
