% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_list_to_df.R
\name{cog_list_to_df}
\alias{cog_list_to_df}
\title{Convertir les donnees du COG d'un type liste a un type dataframe}
\usage{
cog_list_to_df(list)
}
\arguments{
\item{list}{la liste de donnees a convertir}
}
\value{
Renvoie une table de donnees
}
\description{
Convertir les donnees du COG d'un type liste a un type dataframe
}
