% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogiter.R
\docType{package}
\name{COGiter}
\alias{COGiter}
\title{COGiter : Fonctions, données et fonds de carte pour gérer les données au territoire.}
\description{
Ce package permet notamment de gérer les fusions de communes intervenues entre 2003 et 2024.
}
\details{
Pour apprendre comment utiliser COGiter, utiliser les vignettes :
\code{browseVignettes(package = "COGiter")}
}
