% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{communes}
\alias{communes}
\title{Liste des communes du dernier millésime connu.}
\format{A data frame with 35357 rows and 19 variables:
\describe{
\item{DEPCOM}{Code Insee}
\item{NOM_DEPCOM}{Nom de la commune}
\item{EPCI}{Code siren de l'EPCI de rattachement}
\item{NOM_EPCI}{Code siren de l'EPCI de rattachement}
\item{DEP}{Code Insee du département de rattachement}
\item{NOM_DEP}{Nom du département de rattachement}
\item{REG}{Code Insee de la région de rattachement}
\item{NOM_REG}{Nom de la région de rattachement}
\item{REGIONS_DE_L_EPCI}{liste des régions dans la quelle l'epci de rattachement apparait}
}}
\source{
\url{https://www.insee.fr/fr/information/2016807}
}
\usage{
communes
}
\description{
Un dataset contenant pour chaque commune de France actuelle son code Insee, son nom et les identifiants des territoire supra
}
\keyword{datasets}
