% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{communes_geo}
\alias{communes_geo}
\title{Contour des communes métropolitaines pour le dernier millésime connu}
\format{A data frame with 35273 rows and 2 variables:
\describe{
\item{DEPCOM}{Code Insee}
\item{geometry}{géométrie}
\item{AREA}{Aire de la commune}
}}
\usage{
communes_geo
}
\description{
Un dataset contenant pour chaque commune de France métropolitaine actuelle son contour
}
\keyword{datasets}
