% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{communes_geo}
\alias{communes_geo}
\alias{communes_metro_geo}
\alias{communes_971_geo}
\alias{communes_972_geo}
\alias{communes_973_geo}
\alias{communes_974_geo}
\alias{communes_976_geo}
\title{Contour des communes de France pour le dernier millésime connu.}
\format{
Un dataframe avec une ligne par commune et 3 variables.
\describe{
\item{DEPCOM}{Code Insee}
\item{AREA}{Aire de la commune en m2}
\item{geometry}{géométrie}
}
}
\usage{
communes_geo

communes_metro_geo

communes_971_geo

communes_972_geo

communes_973_geo

communes_974_geo

communes_976_geo
}
\description{
Contour des communes de France pour le dernier millésime connu.
}
\details{
Des datasets contenant pour chaque commune actuelle son contour.
Un dataset pour la france métropolitaine, un pour chaque DROM, et un assemblage de ces jeux de données pour pourvoir réaliser des cartes France Métropolitaine et DROM.
}
\keyword{datasets}
