% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departements}
\alias{departements}
\title{Liste des départements du dernier millésime connu.}
\format{A data frame with 101 rows and 6 variables:
\describe{
\item{REG}{Code Insee de la région de rattachement}
\item{DEP}{Code Insee du département}
\item{CHEFLIEU}{Chef-lieu d'arrondissement, de département, de région ou bureau centralisateur}
\item{TNCC}{Type de nom en clair}
\item{NCC}{Nom en clair (majuscules)}
\item{NOM_DEP}{Nom du département}
}}
\source{
\url{https://www.insee.fr/fr/information/3363419#titre-bloc-23}
}
\usage{
departements
}
\description{
Un dataset contenant pour chaque départements de France actuelle son code Insee, son nom et les identifiants des territoire supra
}
\keyword{datasets}
