% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departements_geo}
\alias{departements_geo}
\title{Contour des départements métropolitains pour le dernier millésime connu}
\format{A data frame with 96 rows and 2 variables:
\describe{
\item{DEP}{Code Insee}
\item{geometry}{géométrie}
}}
\usage{
departements_geo

departements_geo
}
\description{
Un dataset contenant pour chaque département de France métropolitaine actuelle son contour

Un dataset contenant pour chaque région  de France métropolitaine actuelle son contour
}
\keyword{datasets}
