% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departements_geo}
\alias{departements_geo}
\alias{departements_metro_geo}
\alias{departements_971_geo}
\alias{departements_972_geo}
\alias{departements_973_geo}
\alias{departements_974_geo}
\alias{departements_976_geo}
\title{Contour des départements pour le dernier millésime connu}
\format{
Un dataframe avec une ligne par departement et 3 variables.
\describe{
\item{DEP}{Code Insee}
\item{geometry}{géométrie}
\item{AREA}{Aire du departement}
}
}
\usage{
departements_geo

departements_metro_geo

departements_971_geo

departements_972_geo

departements_973_geo

departements_974_geo

departements_976_geo
}
\description{
Contour des départements pour le dernier millésime connu
}
\details{
Un dataset contenant pour chaque département de France métropolitaine son contour
Un dataset pour la france métropolitaine, un pour chaque DROM, et un assemblage de ces jeux de données pour pourvoir réaliser des cartes France Métropolitaine et DROM.
}
\keyword{datasets}
