% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epci}
\alias{epci}
\title{Liste des epci du dernier millésime connu.}
\format{
A data frame with 1253 rows and 7 variables:
\describe{
\item{EPCI}{Code siren de l'EPCI}
\item{NOM_EPCI}{Nom de l'EPCI}
\item{DEP_SIEGE_EPCI}{Code du departement de la commune siege de l'EPCI}
\item{NATURE_EPCI}{Type de l'epci : ME,CA,CU,CC}
\item{DEPARTEMENTS_DE_L_EPCI}{Liste des départements dans lesquels se retrouve au moins une commune de l'epci}
\item{REGIONS_DE_L_EPCI}{Liste des régions dans lesquelles se retrouve au moins une commune de l'epci}
}
}
\source{
\url{https://www.insee.fr/fr/information/2510634}
}
\usage{
epci
}
\description{
Un dataset contenant chaque epci de France
}
\keyword{datasets}
