% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epci_geo}
\alias{epci_geo}
\title{Contour des epci métropolitains pour le dernier millésime connu}
\format{A data frame with 1241 rows and 2 variables:
\describe{
\item{EPCI}{Code Sirene}
\item{geometry}{géométrie}
}}
\usage{
epci_geo
}
\description{
Un dataset contenant pour chaque epci de France métropolitaine actuelle son contour
}
\keyword{datasets}
