% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epci_geo_xx}
\alias{epci_geo_xx}
\alias{epci_geo}
\alias{epci_metro_geo}
\alias{epci_971_geo}
\alias{epci_972_geo}
\alias{epci_973_geo}
\alias{epci_974_geo}
\alias{epci_976_geo}
\title{Contour des epci de France pour le dernier millésime connu.}
\format{
Un dataframe avec une ligne par epci et 3 variables.
\describe{
\item{EPCI}{Code Sirene}
\item{AREA}{Aire de l'EPCI en m2}
\item{geometry}{géométrie}
}
}
\usage{
epci_geo

epci_metro_geo

epci_971_geo

epci_972_geo

epci_973_geo

epci_974_geo

epci_976_geo
}
\description{
Contour des epci de France pour le dernier millésime connu.
}
\details{
Des datasets contenant pour chaque epci actuel son contour.
Un dataset pour la france métropolitaine, un pour chaque DROM, et un assemblage de ces jeux de données pour pourvoir réaliser des cartes France Métropolitaine et DROM.
}
\keyword{datasets}
