% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_cog.R
\name{filtrer_cog}
\alias{filtrer_cog}
\title{Filtrer un fichier du cog sur un sous ensemble du territoire}
\usage{
filtrer_cog(
  .data,
  depcom = NULL,
  epci = NULL,
  dep = NULL,
  reg = NULL,
  garder_supra = "non",
  epci_complet = FALSE
)
}
\arguments{
\item{.data}{la table de données a filtrer}

\item{depcom}{la commune sur laquelle filtrer les données}

\item{epci}{l'epci sur lequel filtrer les données}

\item{dep}{le departement sur lequel filtrer les données}

\item{reg}{la region sur laquelle filtrer les données}

\item{garder_supra}{">" si on souhaite garder les territoires supra, ">=" si on souhaite garder les territoires supra et du même niveau que celui sélectionné, sauf si le filtre est opéré sur une commune, auquel cas seuls les territoires supra auquel appartient la commune sont renvoyés. "non" par défaut.}

\item{epci_complet}{booleen, TRUE si on souhaite garder toutes les communes des EPCI du département ou de la région}
}
\value{
Renvoie une table de donnees filtrées
}
\description{
Filtrer un fichier du cog sur un sous ensemble du territoire
}
\details{
Les parametres depcom, epci, dep et reg sont exclusifs les uns des autres.
}
\examples{
 pop2015_cogifiee_Corse <- cogifier(pop2015,  code_commune = DEPCOM, as_df = TRUE) \%>\%
   filtrer_cog(reg = "94", garder_supra = '>')

}
