% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_geo_data_gouv.R
\name{get_communes_departement}
\alias{get_communes_departement}
\title{Obtenir la liste des communes d'un département}
\usage{
get_communes_departement(code_departement)
}
\arguments{
\item{code_departement}{le code département insee}
}
\value{
un dataframe
}
\description{
Obtenir la liste des communes d'un département
}
\examples{
\dontrun{
get_communes_departement('44')
}
}
