% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_geo_data_gouv.R
\name{get_departements_region}
\alias{get_departements_region}
\title{Obtenir la liste des départements d'une région}
\usage{
get_departements_region(code_region)
}
\arguments{
\item{code_region}{code région insee}
}
\value{
un dataframe
}
\description{
Obtenir la liste des départements d'une région
}
\examples{
\dontrun{
get_departements_region('52')
}
}
