% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_geo_data_gouv.R
\name{get_geom_commune}
\alias{get_geom_commune}
\title{Obtenir une commune avec son contour geo de l'api geo.api.gouv.fr}
\usage{
get_geom_commune(code_commune)
}
\arguments{
\item{code_commune}{le code commune insee}
}
\value{
un spatial dataframe
}
\description{
Obtenir une commune avec son contour geo de l'api geo.api.gouv.fr
}
\examples{
\dontrun{
get_geom_communes('44109')
}
}
