% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_geo_data_gouv.R
\name{get_geom_communes_departement}
\alias{get_geom_communes_departement}
\title{Obtenir la liste des contours communaux pour un département}
\usage{
get_geom_communes_departement(code_departement)
}
\arguments{
\item{code_departement}{le code département insee}
}
\value{

}
\description{
Obtenir la liste des contours communaux pour un département
}
\examples{
\dontrun{
geom_communes_d44 <- get_geom_communes_departement('44')
plot(geom_communes_d44)
}
}
