% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_geo_data_gouv.R
\name{get_geom_communes_region}
\alias{get_geom_communes_region}
\title{Obtenir la liste des contours communaux pour une région}
\usage{
get_geom_communes_region(code_region)
}
\arguments{
\item{code_region}{code région insee}
}
\value{

}
\description{
Obtenir la liste des contours communaux pour une région
}
\examples{
\dontrun{
geom_communes_r52 <- get_geom_communes_region('52')
plot(geom_communes_r52)
}
}
