% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_cog_geo.R
\name{get_map}
\alias{get_map}
\title{Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM}
\usage{
get_map(depcom = NULL, epci = NULL, dep = NULL, reg = NULL)
}
\arguments{
\item{depcom}{la commune sur laquelle filtrer les données}

\item{epci}{l'epci sur lequel filtrer les données}

\item{dep}{le departement sur lequel filtrer les données}

\item{reg}{la region sur laquelle filtrer les données}
}
\value{
une liste de spatial dataframe
}
\description{
Obtenir les cartes en fonction de la localisation en métropole ou dans les DROM
}
\examples{
get_map(epci = '244400404')
}
