% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{list_com_in_dep}
\alias{list_com_in_dep}
\title{Renvoie la liste des communes pour un département}
\usage{
list_com_in_dep(departement)
}
\arguments{
\item{departement}{Character. Le code du département concerné}
}
\value{
une liste de caractère
}
\description{
Renvoie la liste des communes pour un département
}
\examples{
list_com_in_dep("24")
}
