% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{list_com_in_epci}
\alias{list_com_in_epci}
\title{Renvoie la liste des communes pour un epci}
\usage{
list_com_in_epci(epci)
}
\arguments{
\item{epci}{Character. Le code de l'epci concerné}
}
\value{
une liste de caractère
}
\description{
Renvoie la liste des communes pour un epci
}
\examples{
list_com_in_epci("200070647")
}
