% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{list_dep_in_reg}
\alias{list_dep_in_reg}
\title{Renvoie la liste des departements pour une region}
\usage{
list_dep_in_reg(region)
}
\arguments{
\item{region}{Character. Le code de la region concernee}
}
\value{
une liste de caractere
}
\description{
Renvoie la liste des departements pour une region
}
\examples{
list_dep_in_reg("75")
}
