% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{list_epci_in_reg}
\alias{list_epci_in_reg}
\title{Renvoie la liste des EPCI pour une région}
\usage{
list_epci_in_reg(regions)
}
\arguments{
\item{regions}{Character. Le code de la région concernée}
}
\value{
une liste de caractère
}
\description{
Renvoie la liste des EPCI pour une région
}
\examples{
list_epci_in_reg("52")
}
