% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{list_xx_in_yy}
\alias{list_xx_in_yy}
\alias{list_dep_in_reg}
\alias{list_epci_in_dep}
\alias{list_epci_in_reg}
\alias{list_com_in_epci}
\alias{list_com_in_dep}
\alias{list_com_in_reg}
\title{Renvoie la liste des territoires pour un ou des territoires}
\usage{
list_dep_in_reg(reg)

list_epci_in_dep(dep)

list_epci_in_reg(reg)

list_com_in_epci(epci)

list_com_in_dep(dep)

list_com_in_reg(reg)
}
\arguments{
\item{reg}{Vecteur de caractère. Le code de la région concernée.}

\item{dep}{Vecteur de caractère. Le code du département concernée.}

\item{epci}{Vecteur de caractère. Le code de l'epci concernée.}
}
\value{
Un vecteur de caractère.
}
\description{
Renvoie la liste des territoires pour un ou des territoires

Renvoie la liste des départements pour une ou des régions

Renvoie la liste des EPCI pour un département

Renvoie la liste des EPCI pour une région

Renvoie la liste des communes pour un epci

Renvoie la liste des communes pour un ou des départements

Renvoie la liste des communes pour une région
}
\examples{
list_dep_in_reg("75")
list_epci_in_dep("24")
list_epci_in_reg("52")
list_com_in_epci("200070647")
list_com_in_dep("24")
list_com_in_reg("52")
}
