% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvtcommunes}
\alias{mvtcommunes}
\title{Table des mouvements des communes}
\format{
A data frame with 13734 rows and 14 variables:
\describe{
\item{MOD}{Type d'événement de communes}
\item{DATE_EFF}{Date d'effet}
\item{TYPECOM_AV}{Type de la commune avant événement}
\item{COM_AV}{Code de la commune avant événement}
\item{TNCC_AV}{Type de nom en clair }
\item{NCC_AV}{Nom en clair (majuscules)}
\item{NCCENR_AV}{Nom en clair (typographie riche)}
\item{LIBELLE_AV}{Nom en clair (typographie riche) avec article}
\item{TYPECOM_AP}{Type de la commune après événement}
\item{COM_AP}{Code de la commune après événement}
\item{TNCC_AP}{Type de nom en clair }
\item{NCC_AP}{Nom en clair (majuscules)}
\item{NCCENR_AP}{Nom en clair (typographie riche)}
\item{LIBELLE_AP}{Nom en clair (typographie riche) avec article}
}
}
\source{
\url{https://www.insee.fr/fr/information/3720946#titre-bloc-23}
}
\usage{
mvtcommunes
}
\description{
Un dataset listant les mouvements opérés sur la carte communale
}
\keyword{datasets}
