% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_zone.R
\name{nom_zone}
\alias{nom_zone}
\title{Obtenir le nom d'une zone}
\usage{
nom_zone(type_zone, code_zone)
}
\arguments{
\item{type_zone}{Type de Zone}

\item{code_zone}{Code de la Zone}
}
\value{
une chaîne de caractère
}
\description{
Obtenir le nom d'une zone
}
\examples{
nom_zone(c("Communes","Communes"),c("75056","44109"))
}
