% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_geo}
\alias{regions_geo}
\title{Contour des régions métropolitaines pour le dernier millésime connu}
\format{A data frame with 13 rows and 2 variables:
\describe{
\item{REG}{Code Insee}
\item{geometry}{géométrie}
}}
\usage{
regions_geo
}
\description{
Un dataset contenant pour chaque région  de France métropolitaine actuelle son contour
}
\keyword{datasets}
