% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_geo_xx}
\alias{regions_geo_xx}
\alias{regions_geo}
\alias{regions_metro_geo}
\alias{regions_971_geo}
\alias{regions_972_geo}
\alias{regions_973_geo}
\alias{regions_974_geo}
\alias{regions_976_geo}
\title{Contour des régions de France pour le dernier millésime connu.}
\format{
Un dataframe avec une ligne par région et 3 variables.
\describe{
\item{REG}{Code Insee}
\item{AREA}{Aire de la region en m2}
\item{geometry}{géométrie}
}
}
\usage{
regions_geo

regions_metro_geo

regions_971_geo

regions_972_geo

regions_973_geo

regions_974_geo

regions_976_geo
}
\description{
Contour des régions de France pour le dernier millésime connu.
}
\details{
Dataset des contours régionaux : un pour la france métropolitaine, un pour chaque DROM, et un assemblage de ces jeux de données pour pourvoir réaliser des cartes France Métropolitaine et DROM.
}
\keyword{datasets}
