% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_passage_com_epci}
\alias{table_passage_com_epci}
\title{Rattachement des communes à leur epci.}
\format{
A data frame with 34955 rows and 2 variables:
\describe{
\item{DEPCOM}{Code Insee de la commune}
\item{EPCI}{Code siren de l'EPCI de rattachement}
}
}
\source{
\url{https://www.insee.fr/fr/information/2510634}
}
\usage{
table_passage_com_epci
}
\description{
Un dataset contenant pour chaque commune de France actuelle le code sirene de son EPCI
}
\keyword{datasets}
