% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_passage_com_historique}
\alias{table_passage_com_historique}
\title{Table de passage des fusions de communes.}
\format{A data frame with 39161 rows and 2 variables:
\describe{
\item{DEPCOM_HIST}{Code Insee de la commune historique}
\item{DEPCOM}{Code Insee de la commune}
}}
\source{
\url{https://www.insee.fr/fr/information/3363419#titre-bloc-3}
}
\usage{
table_passage_com_historique
}
\description{
Un dataset contenant pour chaque commune historique de France le code de la commune actuelle de rattachement
}
\keyword{datasets}
