% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_drom.R
\name{translate_drom}
\alias{translate_drom}
\title{Fonction interne de translation des DROM pour construire la carte france métro et DROM}
\usage{
translate_drom(code_dom, destination, scale = 1, angle = 0, epsg = 2154)
}
\arguments{
\item{code_dom}{code departement du drom}

\item{destination}{x,y du lieu de destination dans l'epsg défini}

\item{scale}{facteur de redimentionnement du territoire}

\item{angle}{facteur de rotation du territoire}

\item{epsg}{code epsg dans lequel effectuer la transformation}
}
\value{
un spatial dataframe
}
\description{
Fonction interne de translation des DROM pour construire la carte france métro et DROM
}
