% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_zone.R
\name{trouver_zone}
\alias{trouver_zone}
\title{Trouver une zone à partir d'une chaîne de caractère}
\usage{
trouver_zone(pattern)
}
\arguments{
\item{pattern}{la chaîne de caractère à identifier}
}
\value{
un tibble
}
\description{
Trouver une zone à partir d'une chaîne de caractère
}
\examples{
trouver_zone(pattern = "Belleville")
}
