% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zonage_pinel_r52}
\alias{zonage_pinel_r52}
\title{Zonage Pinel région Pays de la Loire.}
\format{A data frame with 70 rows and 2 variables:
\describe{
\item{DEPCOM}{Code Insee de la commune}
\item{zonage_pinel}{Zonage Pinel}
}}
\usage{
zonage_pinel_r52
}
\description{
Un dataset contenant pour la liste des communes 2019 éligible au pinel en 2019
}
\keyword{datasets}
