% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_list_to_df.R
\encoding{UTF-8}
\name{zone_list_to_df}
\alias{zone_list_to_df}
\title{Preparer les données pour passer d'un type liste a un type dataframe}
\usage{
zone_list_to_df(.data, typezone)
}
\arguments{
\item{.data}{la table de données a convertir}

\item{typezone}{le type de zonage}
}
\value{
Renvoie une table de données renommée
}
\description{
Preparer les données pour passer d'un type liste a un type dataframe
}
\keyword{internal}
