% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_zonage.R
\name{ajouter_typologie}
\alias{ajouter_typologie}
\title{Ajouter une typologie supra communale spécifique à une table cogifiée}
\usage{
ajouter_typologie(
  .data,
  epci = FALSE,
  departements = TRUE,
  regions = TRUE,
  metro = TRUE,
  metrodrom = TRUE,
  drom = TRUE,
  franceprovince = TRUE,
  zonage_df,
  var_depcom = DEPCOM,
  var_code_zone = CodeZone,
  var_type_zone = TypeZone,
  var_zone = Zone
)
}
\arguments{
\item{.data}{la table de données a filtrer}

\item{epci}{booléen TRUE si on souhaite des données de chaque classe de la typologiee à l'epci}

\item{departements}{booléen TRUE si on souhaite des données de chaque classe de la typologiee  au département}

\item{regions}{booléen TRUE si on souhaite des données de chaque classe de la typologiee  à la région}

\item{metro}{booléen TRUE si on souhaite des données de chaque classe de la typologiee sur la France métropolitaine}

\item{metrodrom}{booléen TRUE si on souhaite des données de chaque classe de la typologiee sur la France métropolitaine et des DROM}

\item{drom}{booléen TRUE si on souhaite des données de chaque classe de la typologiee pour les Départements et régions d'outre mer}

\item{franceprovince}{booléen TRUE si on souhaite des données de chaque classe de la typologiee sur la France de province}

\item{zonage_df}{le dataframe contenant le rattachement entre le code commune et la typologie}

\item{var_depcom}{le nom de la variable code commune dans zonage_df}

\item{var_code_zone}{le nom de la variable code zone dans zonage_df}

\item{var_type_zone}{le nom de la variable type zone dans zonage_df}

\item{var_zone}{le nom de la variable zone dans zonage_df}
}
\value{
Renvoie une table de donnees cogifiée augmentée des calculs pour ces nouvelles typologies
}
\description{
Permet de rajouter à une table COGifiée de nouveaux type de zonage basé sur des agrégations par epci, departements, régions... de typologie de commune (exemple : aire d'attraction des villes de l'insee, zonage ABC pour le logement)
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(COGiter)
zonage_aav <- charger_zonage("AAV2020")
cogpop2015 <- cogifier(pop2015)
cogpop2015aav <- ajouter_typologie(cogpop2015,
                                  zonage_df = zonage_aav)
}
