% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_zonage.R
\name{ajouter_zonage}
\alias{ajouter_zonage}
\title{Ajouter un zonage supra communal spécifique à une table cogifiée}
\usage{
ajouter_zonage(
  .data,
  zonage_df,
  var_depcom = DEPCOM,
  var_code_zone = CodeZone,
  var_type_zone = TypeZone,
  var_zone = Zone
)
}
\arguments{
\item{.data}{la table de données cogifiee, comportant des donnees communales, a partir de laquelle calculer les agregats selon le nouveau zonage souhaite}

\item{zonage_df}{le dataframe contenant le rattachement entre le code commune et le nouveau zonage, contenant 4 informations : le code commune, le code zone dont relève la commune, le libellé de la zone dont relève la commune, le nom du zonage (par exemple 'bassins versants')}

\item{var_depcom}{le nom de la variable code commune dans zonage_df}

\item{var_code_zone}{le nom de la variable désignant le code zone dans zonage_df}

\item{var_type_zone}{le nom de la variable désignant le nom du zonage dans zonage_df}

\item{var_zone}{le nom de la variable désignant le libelle de la zone dans zonage_df}
}
\value{
Renvoie une table de données cogifiée augmentée des agrégats selon ce nouveau zonage
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
