% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_zonage.R
\name{charger_zonage}
\alias{charger_zonage}
\title{Charger un zonage présent dans la table table_passage_communes_zonages}
\usage{
charger_zonage(zonage)
}
\arguments{
\item{zonage}{Caractère - Zonage parmi ceux disponibles}
}
\value{
Un dataframe
}
\description{
Charger un zonage présent dans la table table_passage_communes_zonages
}
\examples{
charger_zonage("ARR")
}
