% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_com_non_a_jour}
\alias{code_com_non_a_jour}
\title{Test si les codes communes sont des codes ayant existé mais qui ne sont plus valide}
\usage{
code_com_non_a_jour(depcom)
}
\arguments{
\item{depcom}{Un vecteur de caractères. Un vecteur de codes commune.}
}
\value{
Un vecteur de booléens.
}
\description{
Test si les codes communes sont des codes ayant existé mais qui ne sont plus valide
}
\examples{
code_com_non_a_jour("75056")
}
