% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_com_plm}
\alias{code_com_plm}
\title{test si les codes communes sont ceux d'un arrondissement de Paris, Lyon ou Marseille}
\usage{
code_com_plm(depcom)
}
\arguments{
\item{depcom}{Character. Un vecteur de codes commune.}
}
\value{
Un vecteur de booléens.
}
\description{
test si les codes communes sont ceux d'un arrondissement de Paris, Lyon ou Marseille
}
\examples{
code_com_plm(c("75056","75112"))
}
