% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_epci_of_com}
\alias{code_epci_of_com}
\title{Renvoie le code de l'epci d appartenance d une commune}
\usage{
code_epci_of_com(depcom)
}
\arguments{
\item{depcom}{Character. Un vecteur de code commune.}
}
\value{
Un vecteur
}
\description{
Renvoie le code de l'epci d appartenance d une commune
}
\examples{
code_epci_of_com(c("24037"))
}
