% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_reg_of_dep}
\alias{code_reg_of_dep}
\title{Renvoie le code de la région d appartenance d un département}
\usage{
code_reg_of_dep(departement)
}
\arguments{
\item{departement}{Character. Le code du département concerné}
}
\value{
un caractere
}
\description{
Renvoie le code de la région d appartenance d un département
}
\examples{
code_reg_of_dep("24")
}
