% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp_territoires.R
\name{code_reg_of_epci}
\alias{code_reg_of_epci}
\title{Renvoie le code de la région d appartenance d un epci}
\usage{
code_reg_of_epci(epci)
}
\arguments{
\item{epci}{Character. Le code de l'epci concerné}
}
\value{
un caractere
}
\description{
Renvoie le code de la région d appartenance d un epci
}
\examples{
code_reg_of_epci("200070647")
}
