% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_df_to_list.R
\encoding{UTF-8}
\name{cog_df_to_list}
\alias{cog_df_to_list}
\title{Convertir les donn\encoding{é}es du COG d'un type dataframe a un type liste}
\usage{
cog_df_to_list(.data)
}
\arguments{
\item{.data}{la liste de donn\encoding{é}es a convertir}
}
\value{
Renvoie une liste de dataframe
}
\description{
Convertir les donn\encoding{é}es du COG d'un type dataframe a un type liste
}
