% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogifier.R
\name{cogifier}
\alias{cogifier}
\title{Consolider une table de données à la commune à tous les échelles du cog}
\usage{
cogifier(.data, code_commune = DEPCOM, metro = T, metrodrom = F)
}
\arguments{
\item{.data}{la table de données à convertire}

\item{code_commune}{le nom de la variable contenant le code commune sur 5 charactères}

\item{metro}{booléen TRUE si on souhaite une aggrégation sur le contour de la France métropolitaine}

\item{metrodrom}{booléen TRUE si on souhaite une aggrégation sur le contour de la France métropolitaine et des DROM}
}
\value{
la fonction renvoie une table de données
}
\description{
Consolider une table de données à la commune à tous les échelles du cog
}
