% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogifier.R
\name{cogifier}
\alias{cogifier}
\title{Consolider une table de donn\encoding{é}es à la commune à tous les \encoding{é}chelles du cog}
\usage{
cogifier(.data, code_commune = DEPCOM, communes = T, epci = T,
  departements = T, regions = T, metro = T, metrodrom = F,
  as_df = T)
}
\arguments{
\item{.data}{la table de donn\encoding{é}es à convertire}

\item{code_commune}{le nom de la variable contenant le code commune sur 5 charactères}

\item{communes}{bool\encoding{é}en TRUE si on souhaite des donn\encoding{é}es à la commune}

\item{epci}{bool\encoding{é}en TRUE si on souhaite des donn\encoding{é}es à l'epci}

\item{departements}{bool\encoding{é}en TRUE si on souhaite des donn\encoding{é}es au d\encoding{é}partement}

\item{regions}{bool\encoding{é}en TRUE si on souhaite des donn\encoding{é}es à la r\encoding{é}gion}

\item{metro}{bool\encoding{é}en TRUE si on souhaite des donn\encoding{é}es France m\encoding{é}tropolitaine}

\item{metrodrom}{bool\encoding{é}en TRUE si on souhaite des donn\encoding{é}es France m\encoding{é}tropolitaine et des DROM}

\item{as_df}{bool\encoding{é}en TRUE si on souhaite des donn\encoding{é}es sous un seul dataframe, false si on souhaite une liste de dataframe par type de zone}
}
\value{
la fonction renvoie un dataframe ou une liste de dataframe
}
\description{
Consolider une table de donn\encoding{é}es à la commune à tous les \encoding{é}chelles du cog
}
