% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogifier.R
\name{cogifier}
\alias{cogifier}
\title{Consolider une table de données à la commune à tous les échelles du cog}
\usage{
cogifier(
  .data,
  code_commune = DEPCOM,
  communes = TRUE,
  epci = TRUE,
  departements = TRUE,
  regions = TRUE,
  metro = TRUE,
  metrodrom = FALSE,
  franceprovince = FALSE,
  drom = FALSE,
  as_df = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{.data}{la table de données à convertir}

\item{code_commune}{le nom de la variable contenant le code commune sur 5 charactères}

\item{communes}{booléen TRUE si on souhaite des données à la commune}

\item{epci}{booléen TRUE si on souhaite des données à l'epci}

\item{departements}{booléen TRUE si on souhaite des données au département}

\item{regions}{booléen TRUE si on souhaite des données à la région}

\item{metro}{booléen TRUE si on souhaite des données France métropolitaine}

\item{metrodrom}{booléen TRUE si on souhaite des données France métropolitaine et des DROM}

\item{franceprovince}{booléen TRUE si on souhaite des données France de province}

\item{drom}{booléen TRUE si on souhaite des données Départements et régions d'outre mer}

\item{as_df}{booléen TRUE si on souhaite des données sous un seul dataframe, FALSE si on souhaite une liste de dataframe par type de zone}

\item{na.rm}{argument(s) passé(s) à la fonction d'aggrégation (sum), na.rm=F par défaut}
}
\value{
Renvoie un dataframe ou une liste de dataframe
}
\description{
Consolider une table de données à la commune à tous les échelles du cog
}
\examples{
pop2015_cogifiee <- cogifier(pop2015, code_commune = DEPCOM)

}
