% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogiter.R
\docType{package}
\name{COGiter-package}
\alias{COGiter}
\alias{COGiter-package}
\title{COGiter : Fonctions, données et fonds de carte pour gérer les données au territoire.}
\description{
Ce package permet notamment de gérer les fusions de communes intervenues entre 2003 et 2025.
}
\details{
Pour apprendre comment utiliser COGiter, utiliser les vignettes :
\code{browseVignettes(package = "COGiter")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maeltheuliere.github.io/COGiter/}
  \item Report bugs at \url{https://github.com/MaelTheuliere/COGiter/issues}
}

}
\author{
\strong{Maintainer}: Juliette ENGELARE-LEFEBVRE \email{juliette.engelaere@developpement-durable.gouv.fr}

Authors:
\itemize{
  \item Maël THEULIERE \email{mael.theuliere@gmail.com}
}

Other contributors:
\itemize{
  \item Caroline COUDRIN \email{caroline.coundrin@developpement-durable.gouv.fr} [contributor]
  \item Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID}) [contributor]
}

}
\keyword{internal}
