% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{communes_cog}
\alias{communes_cog}
\title{Liste des communes du dernier millésime connu.}
\format{A data frame with 37932 rows and 11 variables:
\describe{
\item{TYPECOM}{Code Insee}
\item{DEPCOM}{Code Insee}
\item{NOM_DEPCOM}{Nom en clair (typographie riche) avec article }
\item{REG}{Code Insee de la région de rattachement}
\item{DEP}{Code Insee du département de rattachement}
\item{ARR}{Code arrondissement}
\item{TNCC}{Type de nom en clair}
\item{NCC}{Nom en clair (majuscules)}
\item{NCCENR}{Nom en clair (typographie riche)}
\item{CAN}{Code canton. Pour les communes « multi-cantonales » code décliné de 99 à 90 (pseudo-canton) ou de 89 à 80 (communes nouvelles) }
\item{COMPARENT}{Code de la commune parente pour les arrondissements municipaux et les communes associées ou déléguées. }
}}
\source{
\url{https://www.insee.fr/fr/information/3720946#titre-bloc-3}
}
\usage{
communes_cog
}
\description{
Un dataset contenant chaque commune, communes déléguées,arrondissements, de France
}
\keyword{datasets}
