% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departements}
\alias{departements}
\title{Liste des départements du dernier millésime connu.}
\format{
A data frame with 101 rows and 7 variables:
\describe{
\item{DEP}{Code Insee du département}
\item{REG}{Code Insee de la région de rattachement}
\item{CHEFLIEU}{Chef-lieu d'arrondissement, de département, de région ou bureau centralisateur}
\item{TNCC}{Type de nom en clair}
\item{NCC}{Nom en clair (majuscules)}
\item{NCCENR}{Nom en clair (typographie riche)}
\item{NOM_DEP}{Nom en clair (typographie riche) avec article}
}
}
\source{
\url{https://www.insee.fr/fr/information/3720946#titre-bloc-14}
}
\usage{
departements
}
\description{
Un dataset contenant chaque départements de France
}
\keyword{datasets}
