% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epci}
\alias{epci}
\title{Liste des epci du dernier millésime connu.}
\format{A data frame with 1264 rows and 4 variables:
\describe{
\item{EPCI}{Code siren de l'EPCI}
\item{NOM_EPCI}{Nom de l'EPCI}
\item{DEPARTEMENTS_DE_L_EPCI}{Liste des départements dans lesquels  se retrouve au moins une commune de l'epci}
\item{REGIONS_DE_L_EPCI}{Liste des régions dans lesquelles  se retrouve au moins une commune de l'epci}
}}
\source{
\url{https://www.insee.fr/fr/information/2510634}
}
\usage{
epci
}
\description{
Un dataset contenant pour chaque epci de France actuel son code Insee, son nom et les identifiants des territoire supra
}
\keyword{datasets}
