% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_cog.R
\encoding{UTF-8}
\name{filtrer_cog}
\alias{filtrer_cog}
\title{Filtrer une fichier du cog sur un sous ensemble du territoire}
\usage{
filtrer_cog(.data, depcom = NULL, epci = NULL, dep = NULL,
  reg = NULL, garder_supra = "non")
}
\arguments{
\item{.data}{la table de donn\encoding{é}es a filtrer}

\item{depcom}{la commune sur laquelle filtrer les donn\encoding{é}es}

\item{epci}{l'epci sur lequel filtrer les donn\encoding{é}es}

\item{dep}{le departement sur lequel filtrer les donn\encoding{é}es}

\item{reg}{la region sur laquelle filtrer les donn\encoding{é}es}

\item{garder_supra}{">" si on souhaite garder les territoires supra, ">=" si on souhaite garder les territoires suppra et du même niveau que celui s\encoding{é}lectionn\encoding{é}}
}
\value{
Renvoie une table de donnees filtrées
}
\description{
Filtrer une fichier du cog sur un sous ensemble du territoire
}
