% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_cog_geo.R
\name{filtrer_cog_geo}
\alias{filtrer_cog_geo}
\title{Filtrer les fonds de carte sur un territoire métropolitain}
\usage{
filtrer_cog_geo(
  depcom = NULL,
  epci = NULL,
  dep = NULL,
  reg = NULL,
  garder_supra = FALSE,
  lg_buffer = 0
)
}
\arguments{
\item{depcom}{la commune sur laquelle filtrer les données}

\item{epci}{l'epci sur lequel filtrer les données}

\item{dep}{le departement sur lequel filtrer les données}

\item{reg}{la region sur laquelle filtrer les données}

\item{garder_supra}{TRUE si on souhaite une carte centrée sur le territoire mais avec les territoires environnant visible, non si on souhaite garder que le territoire sélectionné}

\item{lg_buffer}{si garder_supra est a true, la largeur du buffer a garder autour du territoire (metres)}
}
\value{
une liste de spatial dataframe
}
\description{
Filtrer les fonds de carte sur un territoire métropolitain
}
\examples{
nantes_metropole <- filtrer_cog_geo(epci = '244400404')
plot(nantes_metropole$communes)
}
