% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liste_zone}
\alias{liste_zone}
\title{Une table permettant de lister pour chaque département ou région les zonages infra}
\format{
A data frame with 36338 rows and 7 variables:
\describe{
\item{TypeZone}{Type de la zone : communes,epci,départements,régions}
\item{CodeZone}{Code de la zone}
\item{Zone}{Libellé de la zone}
\item{EPCI}{liste des EPCI ayant cette zone en territoire infra}
\item{NATURE_EPCI}{nature de l'Epci pour les EPCI}
\item{DEP}{liste des départements ayant cette zone en territoire infra}
\item{REG}{liste des régions ayant cette zone en territoire infra}
}
}
\usage{
liste_zone
}
\description{
Une table permettant de lister pour chaque département ou région les zonages infra
}
\keyword{datasets}
