% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Liste des régions du dernier millésime connu.}
\format{A data frame with 18 rows and 5 variables:
\describe{
\item{REG}{Code Insee de la région}
\item{CHEFLIEU}{Chef-lieu d'arrondissement, de département, de région ou bureau centralisateur}
\item{TNCC}{Type de nom en clair}
\item{NCC}{Nom en clair (majuscules)}
\item{NOM_REG}{Nom de la région}
}}
\source{
\url{https://www.insee.fr/fr/information/3363419#titre-bloc-26}
}
\usage{
regions
}
\description{
Un dataset contenant pour chaque régions de France actuelle son code Insee, son nom
}
\keyword{datasets}
