% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc-table_passage_communes_zonages.R
\docType{data}
\name{table_passage_communes_zonages}
\alias{table_passage_communes_zonages}
\title{table_passage_communes_zonages}
\format{
A data frame with 34935 rows and 27 variables:
\describe{
\item{ DEPCOM }{  factor Code commune INSEE}
\item{ ARR }{  factor Code arrondissement }
\item{ LIB_ARR }{  factor Libellé arrondissement}
\item{ CV }{  factor Code canton ville }
\item{ LIB_CV }{  factor Libellé canton ville}
\item{ ZE2020 }{  factor Code zone d'emploi 2020 }
\item{ LIB_ZE2020 }{  factor Libellé zone d'emploi 2020}
\item{ UU2020 }{  factor Code unité urbaine 2020 }
\item{ LIB_UU2020 }{  factor Libellé unité urbaine 2020}
\item{ COMUU2020 }{  factor Statut de la commune dans l'unité urbaine 2020 }
\item{ LIB_COMUU2020 }{  factor Libellé du statut de la commune dans l'unité urbaine 2020}
\item{ TYPUU2020 }{  factor Type d'unité urbaine}
\item{ LIB_TYPUU2020 }{  factor Libellé }
\item{ TUU2017 }{  factor Code tranche d'unité urbaine 2020 calculée sur la population 2017}
\item{ LIB_TUU2017 }{  factor Libellé tranche d'unité urbaine 2020 calculée sur la population 2017}
\item{ TDUU2017 }{  factor Code tranche d'unité urbaine calculée sur la population 2017}
\item{ LIB_TDUU2017 }{  factor Libellé tranche détaillée d'unité urbaine 2020 calculée sur la population 2017}
\item{ AAV2020 }{  factor Code aire d'attraction des villes 2020}
\item{ LIB_AAV2020 }{  factor Libellé aire d'attraction des villes 2020}
\item{ TAAV2017 }{  factor Code tranche d'aire d'attraction des villes 2020 calculée sur la population 2017}
\item{ LIB_TAAV2017 }{  factor Libellé tranche d'aire d'attraction des villes 2020 calculée sur la population 2017}
\item{ TDAAV2017 }{  factor Code tranche détaillée d'aire d'attraction des villes 2020 calculée sur la population 2017}
\item{ LIB_TDAAV2017 }{  factor Libellé tranche détaillée d'aire d'attraction des villes 2020 calculée sur la population 2017}
\item{ CATEAAV2020 }{  factor Code catégorie commune dans aire d'attraction des villes 2020}
\item{ LIB_CATEAAV2020 }{  factor Libellé catégorie commune dans aire d'attraction des villes 2020}
\item{ BV2022 }{  factor Code Code du bassin de vie 2022}
\item{ LIB_BV2022 }{  factor Libellé Bassin de vie 2022}
}
}
\source{
Insee : https://www.insee.fr/fr/information/2028028
}
\usage{
table_passage_communes_zonages
}
\description{
Table de passage des communes vers des zonages supra.
}
\keyword{datasets}
