% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zonage_abc}
\alias{zonage_abc}
\title{Zonage ABC France entière}
\format{A data frame with 34970 rows and 2 variables:
\describe{
\item{DEPCOM}{Code Insee de la commune}
\item{zonage_abc}{Zonage ABC}
}}
\usage{
zonage_abc
}
\description{
Un dataset contenant pour chaque commune du GOG 2019 le zonage ABC correspondant
}
\keyword{datasets}
