% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zonage_abc_r52}
\alias{zonage_abc_r52}
\title{Zonage ABC région Pays de la Loire.}
\format{A data frame with 1238 rows and 2 variables:
\describe{
\item{DEPCOM}{Code Insee de la commune}
\item{zonage_abc}{Zonage ABC}
}}
\usage{
zonage_abc_r52
}
\description{
Un dataset contenant pour chaque commune du GOG 2019 de la région Pays de la Loire le zonage ABC correspondant
}
\keyword{datasets}
