% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zonage_pinel_r52}
\alias{zonage_pinel_r52}
\title{Zonage Pinel région Pays de la Loire.}
\format{A data frame with 1272 rows and 2 variables:
\describe{
\item{DEPCOM}{Code Insee de la commune}
\item{zonage_pinel}{Zonage Pinel}
}}
\usage{
zonage_pinel_r52
}
\description{
Un dataset contenant pour chaque commune du COG 2018 de la région Pays de la Loire le zonage Pinel correspondant
}
\keyword{datasets}
