% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_df_to_list.R
\encoding{UTF-8}
\name{zone_df_to_list}
\alias{zone_df_to_list}
\title{Préparer les données pour passer d'un type dataframe a un type liste}
\usage{
zone_df_to_list(.data, typezone)
}
\arguments{
\item{.data}{la table de données a convertir}

\item{typezone}{le type de zonage}
}
\value{
Renvoie une table de données renommée
}
\description{
Préparer les données pour passer d'un type dataframe a un type liste
}
\keyword{internal}
