% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_list_to_df.R
\name{zone_list_to_df}
\alias{zone_list_to_df}
\title{Preparer les donnees pour passer d'un type liste a un type dataframe}
\usage{
zone_list_to_df(.data, typezone)
}
\arguments{
\item{.data}{la table de donnees a convertir}

\item{typezone}{le type de zonage}
}
\value{
la fonction renvoie une table de donnees renommee
}
\description{
Preparer les donnees pour passer d'un type liste a un type dataframe
}
